+++ src/main/kotlin/in/kkkev/jjidea/ui/common/JujutsuIcons.kt
@@ -49,12 +49,4 @@ object JujutsuIcons {
 
     @JvmField
     val Repo: Icon = strokeFillIcon("/icons/repo.svg", GRAY)
-
-    // TODO How do we call this for a colored { ... } icon?
-    fun repo(accent: Color) = load("/icons/repo.svg").recolored(
-        mapOf(
-            "accent" to ("stroke" to accent),
-            "accent-fill" to ("fill" to accent)
-        )
-    )
 }